1. Modular package management. 2. Login screen mplemented.
