+++ reactnative_mobile/react_app_todo_list/package.json
@@ -21,9 +21,11 @@
     "react-native": "0.72.5",
     "react-native-gesture-handler": "^2.12.0",
     "react-native-get-random-values": "~1.9.0",
+    "react-native-paper": "^5.11.1",
     "react-native-restart": "^0.0.27",
     "react-native-uuid": "^2.0.1",
     "react-native-webview": "^13.2.2",
+    "react-navigation-stack": "^2.10.4",
     "statsig-react": "^1.30.3",
     "statsig-react-native-expo": "^4.6.1",
     "uuid": "^9.0.1"
