+++ reactnative_mobile/react_app_todo_list/src/Navigation.tsx
@@ -1,0 +1,3 @@
+export type Navigation = {
+  navigate: (scene: string) => void;
+};
