+++ reactnative_mobile/react_app_todo_list/src/assets/delete_icon.png
