+++ reactnative_mobile/react_app_todo_list/src/assets/refresh.png
