+++ reactnative_mobile/react_app_todo_list/src/core/themes.tsx
@@ -1,0 +1,11 @@
+import { DefaultTheme } from "react-native-paper";
+
+export const theme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    primary: "#55BCF6",
+    secondary: "#414757",
+    error: "#f13a59",
+  },
+};
